♻️ ClearEventsButton, ToggleRecordButton을 feature 레이어로 이동시키고 정리한다
