--- src/devtools/models/recording-store.ts
@@ -1,13 +1,0 @@
-import { create } from 'zustand';
-
-const useRecordingStore = create<TRecordingState>(set => ({
-  isRecording: false,
-  setIsRecording: isRecording => set({ isRecording }),
-}));
-
-export { useRecordingStore };
-
-type TRecordingState = {
-  isRecording: boolean;
-  setIsRecording: (isRecording: boolean) => void;
-};
