+++ src/devtools/ui/maximized-view.tsx
@@ -1,22 +1,20 @@
-import { Ban } from 'lucide-react';
 import { useState } from 'react';
-import { useApiRecorder } from '@/features/record';
+import { ClearEventsButton, ToggleRecordButton, useRecordingStore } from '@/features/record';
 import { cn } from '@/shared/lib';
-import { Button, Input, ResizableFrame, ScrollArea, Tabs, TabsContent, TabsList, TabsTrigger } from '@/shared/ui';
+import { Input, ResizableFrame, ScrollArea, Tabs, TabsContent, TabsList, TabsTrigger } from '@/shared/ui';
 import { useDevtoolsViewStore } from '../models/devtools-view-store';
 import type { TTab } from '../types';
-import { RecordControllers } from './record-controllers';
 import { SizeControllers } from './size-controllers';
 
 const MaximizedView = () => {
   const { tab: selectedTab, setTab } = useDevtoolsViewStore();
   const [searchValue, setSearchValue] = useState('');
-  const { events, clear } = useApiRecorder();
+  const { events } = useRecordingStore();
 
   return (
     <ResizableFrame>
       <div className="flex items-center justify-between border-b border-gray-700 px-4 pb-3 text-sm">
-        <RecordControllers />
+        <ToggleRecordButton />
         <SizeControllers buttons={['minimize', 'close']} />
       </div>
       <div className="py-3 px-4 flex-1 h-full">
@@ -52,14 +50,7 @@ const MaximizedView = () => {
             onChange={e => setSearchValue(e.target.value)}
             className="text-xs bg-gray-800 border-gray-600 text-white w-36"
           />
-          <Button
-            variant="ghost"
-            size="icon"
-            onClick={clear}
-            className="ml-auto hover:bg-white/30 rounded-full hover:text-white h-7 w-7"
-          >
-            <Ban />
-          </Button>
+          <ClearEventsButton />
         </div>
 
         <div className="overflow-hidden h-full">
