+++ src/features/record/index.ts
@@ -1,1 +1,4 @@
-export * from './hooks/use-api-recorder';
+export * from './models/recording-store';
+
+export * from './ui/clear-events-button';
+export * from './ui/toggle-record-button';
